Refactor project structure, fix API models, update tests, add README

+++ src/test/java/ru/yandex/praktikum/user/UserEditTest.java
@@ -1,6 +1,6 @@
-package ru.yandex.praktikum.User;
+package ru.yandex.praktikum.user;
 
-import ru.yandex.praktikum.service.Service;
+import io.qameta.allure.Description;
 import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
@@ -8,18 +8,18 @@ import org.apache.http.HttpStatus;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import ru.yandex.praktikum.user.UserAPI;
-import ru.yandex.praktikum.user.UserCredentials;
+import ru.yandex.praktikum.service.Service;
 
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
 
 public class UserEditTest {
 
-    UserAPI userAPI = new UserAPI();
-    String authToken;
+    private final UserAPI userAPI = new UserAPI();
+    private String authToken;
 
     @Before
-    @Step("Начало - создание пользователя")
+    @Step("Начало - создание пользователя и настройка спецификаций")
     public void setUp() {
         Service.setupSpecification();
         Response loginResponse = userAPI.create(UserCredentials.user);
@@ -29,23 +29,27 @@ public class UserEditTest {
     @After
     @Step("Завершение - удаляем созданного пользователя")
     public void tearDown() {
-        userAPI.delete(authToken, UserCredentials.user);
+        if (authToken != null) {
+            userAPI.delete(authToken, UserCredentials.user);
+        }
     }
 
     @Test
     @DisplayName("Изменение данных пользователя (корректная авторизация)")
+    @Description("Проверяем, что авторизованный пользователь может изменить свои данные")
     public void editDataWithAuthorization() {
         userAPI.loginAsAuthResponse(UserCredentials.from(UserCredentials.user));
         Response response = userAPI.editData(authToken, UserCredentials.newUser);
         response.then()
                 .assertThat()
-                .body("success", is( true))
-                .body ("user.name", equalTo(UserCredentials.newFakeName))
-                .body ("user.email", equalTo(UserCredentials.newFakeEmail));
+                .body("success", is(true))
+                .body("user.name", equalTo(UserCredentials.newFakeName))
+                .body("user.email", equalTo(UserCredentials.newFakeEmail));
     }
 
     @Test
     @DisplayName("Изменение данных пользователя (без авторизации)")
+    @Description("Проверяем, что без токена изменить данные пользователя нельзя")
     public void editDataWithoutAuthorization() {
         Response response = userAPI.editDataWithoutToken(UserCredentials.newUser);
         response.then()
@@ -55,5 +59,4 @@ public class UserEditTest {
                 .body("success", is(false))
                 .body("message", equalTo("You should be authorised"));
     }
-
 }
